pyproject.toml
[tool.poetry]
name = "git_run_back"
version = "0.1.0"
description = ""
authors = ["Saken Myrzabekov <myrzabekovsaken89@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
